fix rubocop offenses

+++ app/domain/images/services/create.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Images
   module Services
     class Create
@@ -8,29 +10,22 @@ module Images
       attr_accessor :params, :entity
 
       def call
-        @entity = Image.create(thought: params[:thought])
+        create_entity
 
-        if entity.errors.any?
-          result[:errors] = entity.errors.full_messages
+        return result if result[:errors].present?
 
-          return result
-        end
-
-        filename = ActiveStorage::Filename.new(params[:file][:filename]).sanitized
         file = Tempfile.new(filename, binmode: true)
         begin
-          decode_base64_content = Base64.decode64(params[:file][:file_base64])
-          file.write(decode_base64_content)
+          file.write(decoded_base64_content)
           file.rewind
-          entity.file.attach(io: File.open(file.path),
-                             content_type: params[:file][:type],
-                             filename: filename)
+          attach_file_to_entity(file)
 
           publish_image_event if file_ocr_required?
+
           result[:entity] = entity
         ensure
-           file.close
-           file.unlink
+          file.close
+          file.unlink
         end
 
         result
@@ -42,10 +37,30 @@ module Images
         @result ||= { errors: [], entity: nil }
       end
 
+      def create_entity
+        @entity = Image.create(thought: params[:thought])
+
+        result[:errors] = entity.errors.full_messages if entity.errors.any?
+      end
+
+      def filename
+        @filename ||= ActiveStorage::Filename.new(params[:file][:filename]).sanitized
+      end
+
+      def attach_file_to_entity(file)
+        entity.file.attach(io: File.open(file.path),
+                           content_type: params[:file][:type],
+                           filename: filename)
+      end
+
+      def decoded_base64_content
+        Base64.decode64(params[:file][:file_base64])
+      end
+
       def publish_image_event
         Karafka.producer
                .produce_sync(key: entity.id.to_s,
-                             topic: "text_image.created",
+                             topic: 'text_image.created',
                              payload: { file_path: entity.path,
                                         filename: params[:file][:filename] }.to_json)
       end
